requirements.txt
python-telegram-bot[job-queue]==20.7
qrcode[pil]==7.4.2
Pillow==9.5.0
mercadopago==2.2.0
aiohttp==3.9.1
cryptography==41.0.7
python-dateutil==2.8.2
requests==2.31.0
Flask==3.0.0
flask-socketio==5.3.6
eventlet==0.33.3
python-engineio==4.8.0
python-socketio==5.10.0
mysql-connector-python==8.2.0
pymysql==1.1.0
moviepy==1.0.3
imageio-ffmpeg==0.4.9
openpyxl==3.1.2
